Refactor AddressTest to PersonAddressTest
